requirements.txt
aiofiles==23.2.1
aiogram==3.17.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.1.1
Deprecated==1.2.14
diff-match-patch==20241021
distro==1.9.0
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.6
django-cache-url==3.4.5
django-celery-beat==2.7.0
django-ckeditor==6.7.1
django-debug-toolbar==5.0.1
django-import-export==4.3.4
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-solo==2.4.0
django-timezone-field==7.1
django-tinymce==4.1.0
environs==11.0.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.14
frozenlist==1.4.1
gevent==24.11.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
greenlet==3.1.1
gspread==6.1.2
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.7
ipython==8.31.0
jedi==0.19.2
kombu==5.4.2
lxml==5.3.0
magic-filter==1.0.12
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
netaddr==1.3.0
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
tablib==3.7.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
wrapt==1.16.0
yarl==1.9.4
yookassa==3.3.0
zope.event==5.0
zope.interface==7.2
